Refactor : 콘서트 리팩토링
